Enh #7 - Break down edit form
